Title: 커지는 간선의 값, Time: 169ms, Memory: 5MB, Status: Wrong Answer - CodeTree

+++ 240329/커지는 간선의 값/growing-edge-value.cpp
@@ -63,7 +63,7 @@ int main() {
 
 
     }
-    int sum=0;
+    long long sum=0;
     for(int i=0;i<n-1;i++){
         sum+=k*i;
     }
